+++ tests/test_needtable.py
@@ -7,10 +7,36 @@ from sphinx_testing import with_app
 
 @with_app(buildername="html", srcdir="doc_test/doc_needtable")
 def test_doc_build_html(app, status, warning):
+    import sphinx
+
     app.build()
     html = Path(app.outdir, "index.html").read_text()
     assert "SP_TOO_001" in html
     assert 'id="needtable-index-0"' in html
+
+    # check table caption exists
+    assert "Test table caption" in html
+
+    # check table cpation and table wrapped into figure
+    if sphinx.version_info[0] >= 4:
+        figure = """
+<figure class="align-default" id="needtable-index-0">
+<figcaption>
+<p><span class="caption-text">Test table caption<table class="NEEDS_TABLE rtd-exclude-wy-table docutils align-default">
+"""
+    else:
+        figure = """
+<figure id="needtable-index-0">
+<figcaption>
+<p><span class="caption-text">Test table caption<table class="NEEDS_TABLE rtd-exclude-wy-table docutils">
+"""
+    assert figure in html
+
+    # negative test to check table without caption
+    if sphinx.version_info[0] >= 4:
+        assert '<table class="NEEDS_TABLE rtd-exclude-wy-table docutils align-default" id="needtable-index-1">' in html
+    else:
+        assert '<table class="NEEDS_TABLE rtd-exclude-wy-table docutils" id="needtable-index-1">' in html
 
 
 @with_app(buildername="html", srcdir="doc_test/doc_needtable")