+++ docs/changelog.rst
@@ -29,6 +29,8 @@ License
   (`#305 <https://github.com/useblocks/sphinxcontrib-needs/issues/305>`_)
 * Improvement: Supporting user defined filter code function for :ref:`needs_warnings`
   (`#345 <https://github.com/useblocks/sphinxcontrib-needs/issues/345>`_)
+* Improvement: Supporting caption for :ref:`needtable`
+  (`#348 <https://github.com/useblocks/sphinxcontrib-needs/issues/348>`_)
 
 0.7.1
 -----